+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix submission of the solution due to using old challenge interface
+
 ## 0.2.0 - 30.11.2024
 ### Added
 - Extend cli with possibility to specify directories for: templates, challenges solutions, and data